requirements.txt
django
djangorestframework
slack_sdk
slackeventsapi
python-dotenv
pandas
matplotlib
seaborn
requests